Backend/requirements.txt
flask
flask-wtf
flask-pymongo
flask-login
flask-bcrypt
python-dotenv